Still thinking through valid_moves
I think I'm overcomplicating this. The knight should really build its
own graph given a starting node. So we start at x, add a neighbor to
next valid moves, then recurse again? In the end, we should end with
an entire graph where all the node/vertex neighours are generated.

So, once we have this graph, we can just pass it to the "knight_moves"
method, which is will be an implementation of BFS to find the shortest
path.

Seems like the best way to handle this, is to actually ignore the board
and make a new graph.