Neighbor approach not working
Seems to be an implementation problem. As the graph building process
continues, each node's neighbour instance variable gets bigger and in
is especially bad when there's a reference to an already highly con-
nected node.

I believe this is an implementation problem, as @neighbours is just a
list of references, but it seems to be really taxing Ruby. As I get to
d2, the reference is so deeply nested that it ends up stalling and Ruby
kills the process.

I think I'll switch to the adjacency list approach inside board and see
if that goes any better.